Refactor: 테스트 커버리지 측정용으로 모든 테스트 코드 한 곳으로 이전
